Move distance_to_history into pluigins, remove duplicate time_ file
