+++ zmon_worker_monitor/zmon_worker/functions/__init__.py
@@ -1,13 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from distance_to_history import DistanceWrapper
 from sql import SqlWrapper
 from time_ import TimeWrapper
 from zmon import ZmonWrapper
 
 __all__ = [
-    'DistanceWrapper',
     'SqlWrapper',
     'TimeWrapper',
     'ZmonWrapper'
